Extract OptimizeTableStatement to reduce duplicate logic for MySQL and doris
