+++ parser/sql/statement/src/main/java/org/apache/shardingsphere/sql/parser/sql/common/statement/dal/OptimizeTableStatement.java
@@ -15,13 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal;
+package org.apache.shardingsphere.sql.parser.sql.common.statement.dal;
 
-import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.OptimizeTableStatement;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.MySQLStatement;
+import lombok.Getter;
+import org.apache.shardingsphere.sql.parser.sql.common.segment.generic.table.SimpleTableSegment;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.AbstractSQLStatement;
+
+import java.util.Collection;
+import java.util.LinkedList;
 
 /**
- * MySQL optimize table statement.
+ * Optimize table statement.
  */
-public final class MySQLOptimizeTableStatement extends OptimizeTableStatement implements MySQLStatement {
+@Getter
+public abstract class OptimizeTableStatement extends AbstractSQLStatement implements DALStatement {
+    
+    private final Collection<SimpleTableSegment> tables = new LinkedList<>();
 }
